Initial Commit
This is the first commit for the M2North assesment.

Right now it only displays a typeahead box with US town and state names
which will be used to select the location for the weather broadcast.

States with zip codes are loaded via AJAX and the data was originally
sourced from http://federalgovernmentzipcodes.us/ with some processing
done to get it into a JSON formatted file that the library likes.

The wsdlmethods.php simply shows some interesting information about the
SOAP api with some nice formatting provided by the krumo library sourced
from http://kaloyan.info/krumo/index.php