Added the forecast PHP file
The forecast php file will be the API end point used by the JavaScript
on the page.

It does not do much, it takes a zip code and queries the weather info's
SOAP API and translates the result to a JSON object that's easier to
manage in JavaScript. It also curcumvents the cross domain request
(CORS) problems that would pop up if we tried to query the SOAP API
directly with JavaScript.

In adition to this, it also gives nice user readible results if you
access the page directly through the browser.